handle scenario of expired session and mismatch
Sessions on AOI expire after 10 minutes of inactivity. Therefore if a vote comes
in for an appearance more than 10 minutes after the appearance was first shown,
then the appearance and vote will be from different visitors.

Imagine a situation where a user votes, leaves the computer, and another user
comes to the computer 10 minutes later and starts voting. We treat these as
separate visitors. As such, we are creating a copy of the appearance, so that
the original user's appearance is left unanswered, and the new user's appearance
is a copy of the first appearance and *is* answered.