handle old_visitor_identifier parameter in answers
When either a vote or skip gets passed an old_visitor_identifier parameter,
we want to ensure it matches the original appearance and then create a clone
of the original appearance. The cloned appearance gets the new voter_id
information. The end result is the old session with the original appearance
will have a "Stopped_Voting" and the new session will have a vote + appearance.

[#66601646]