pass along algorithm metadata to prompt cache
For the catchup algorithm, we want to log the metadata about the algorithm used
to select a given prompt (tau and alpha). Because the prompts are selected ahead
of time (before they are shown to a user) and cached, then we need to pass these
details along into the cache. Where we previously only stored the prompt id in
the redis cache, we're now storing a JSON string of a hash containing the keys
id and algorithm.